+++ src/fibra.rs
@@ -57,7 +57,7 @@ impl Fibra {
             async move {
                 Ok::<_, Infallible>(service_fn(move |req| {
                     let appself = appself.clone();
-                    let context = Context::new(appself.clone(), req, address.0, address.1);
+                    let context = Context::new(appself.clone(), address.0, address.1, req);
 
                     async move { appself.handle(context).await }
                 }))