+++ src/route/storage.rs
@@ -1,3 +1,0 @@
-use crate::types::*;
-
-pub type Storage = HashMap<String, bool>;