+++ src/types.rs
@@ -5,7 +5,6 @@ pub(crate) use std::slice::Iter;
 pub(crate) use std::future::Future;
 pub(crate) use std::net::SocketAddr;
 pub(crate) use std::convert::Infallible;
-pub(crate) use std::collections::HashMap;
 pub(crate) use std::panic::AssertUnwindSafe;
 pub(crate) use std::net::TcpListener as StdTcpListener;